refactor(backend): migrate auth to inversify DI module
