+++ backend/container.ts
@@ -1,7 +1,9 @@
+// backend/container.ts
 import 'reflect-metadata';
 import { Container } from 'inversify';
 import prisma from './prismaClient.js';
 import type { PrismaClient } from '@prisma/client';
+import { authModule } from './modules/auth/auth.module.js';
 
 export const PRISMA_CLIENT = Symbol.for('PrismaClient');
 
@@ -9,4 +11,4 @@ export const container = new Container({ defaultScope: 'Singleton' });
 
 container.bind<PrismaClient>(PRISMA_CLIENT).toConstantValue(prisma);
 
-// Module bindings will be added in Steps 2/3/4 via container.load(...)
+container.load(authModule);
