+++ backend/modules/auth/auth.controller.ts
@@ -1,4 +1,4 @@
-// backend/controllers/AuthController.ts
+// backend/modules/auth/auth.controller.ts
 import {
   Controller,
   Route,
@@ -9,18 +9,24 @@ import {
   Tags,
 } from '@tsoa/runtime';
 import HttpStatus from 'http-status';
-import * as authService from '../services/auth.js';
+import { inject, injectable } from 'inversify';
+import { AuthService } from './auth.service.js';
 import type {
   GitHubAuthStatus,
   SwitchAccountBody,
   ErrorResponse,
-} from '../types/index.js';
+} from '../../types/index.js';
 
 export type { GitHubAuthStatus, SwitchAccountBody };
 
+@injectable()
 @Route('api/auth')
 @Tags('Auth')
 export class AuthController extends Controller {
+  constructor(@inject(AuthService) private readonly authService: AuthService) {
+    super();
+  }
+
   /**
    * Get current GitHub CLI authentication status
    */
@@ -30,7 +36,7 @@ export class AuthController extends Controller {
     'GitHub CLI unavailable'
   )
   public async getGitHubStatus(): Promise<GitHubAuthStatus> {
-    return authService.getStatus();
+    return this.authService.getStatus();
   }
 
   /**
@@ -45,6 +51,6 @@ export class AuthController extends Controller {
   public async switchGitHubAccount(
     @Body() body: SwitchAccountBody
   ): Promise<GitHubAuthStatus> {
-    return authService.switchAccount(body.username);
+    return this.authService.switchAccount(body.username);
   }
 }
