+++ backend/modules/auth/auth.service.test.ts
@@ -1,3 +1,4 @@
+// backend/modules/auth/auth.service.test.ts
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
 const mockExecFileAsync = vi.hoisted(() => vi.fn());
@@ -6,11 +7,14 @@ vi.mock('node:util', () => ({
   promisify: () => mockExecFileAsync,
 }));
 
-const { getStatus, switchAccount } = await import('./auth.js');
+const { AuthService } = await import('./auth.service.js');
 
-describe('auth service', () => {
+describe('AuthService', () => {
+  let service: InstanceType<typeof AuthService>;
+
   beforeEach(() => {
     vi.clearAllMocks();
+    service = new AuthService();
   });
 
   describe('getStatus', () => {
@@ -35,7 +39,7 @@ describe('auth service', () => {
         stderr: singleAccountOutput,
       });
 
-      const result = await getStatus();
+      const result = await service.getStatus();
 
       expect(result).toEqual({
         activeAccount: 'octocat',
@@ -49,7 +53,7 @@ describe('auth service', () => {
         stderr: multiAccountOutput,
       });
 
-      const result = await getStatus();
+      const result = await service.getStatus();
 
       expect(result).toEqual({
         activeAccount: 'octocat',
@@ -66,7 +70,7 @@ describe('auth service', () => {
         stderr: '',
       });
 
-      const result = await getStatus();
+      const result = await service.getStatus();
 
       expect(result).toEqual({
         activeAccount: 'octocat',
@@ -79,7 +83,7 @@ describe('auth service', () => {
         stderr: singleAccountOutput,
       });
 
-      const result = await getStatus();
+      const result = await service.getStatus();
 
       expect(result).toEqual({
         activeAccount: 'octocat',
@@ -90,7 +94,7 @@ describe('auth service', () => {
     it('should throw SERVICE_UNAVAILABLE when gh CLI is not available', async () => {
       mockExecFileAsync.mockRejectedValue(new Error('ENOENT'));
 
-      await expect(getStatus()).rejects.toMatchObject({
+      await expect(service.getStatus()).rejects.toMatchObject({
         message: 'GitHub CLI is not available',
         statusCode: 503,
       });
@@ -102,7 +106,7 @@ describe('auth service', () => {
         stderr: 'some unrecognized output',
       });
 
-      await expect(getStatus()).rejects.toMatchObject({
+      await expect(service.getStatus()).rejects.toMatchObject({
         message:
           'No GitHub accounts found. Run "gh auth login" in your terminal.',
         statusCode: 503,
@@ -124,7 +128,7 @@ describe('auth service', () => {
         stderr: noActiveOutput,
       });
 
-      const result = await getStatus();
+      const result = await service.getStatus();
 
       expect(result.activeAccount).toBe('octocat');
     });
@@ -139,10 +143,10 @@ describe('auth service', () => {
 
     it('should switch account and return new status', async () => {
       mockExecFileAsync
-        .mockResolvedValueOnce({ stdout: '', stderr: '' }) // switch command
-        .mockResolvedValueOnce({ stdout: '', stderr: statusAfterSwitch }); // getStatus call
+        .mockResolvedValueOnce({ stdout: '', stderr: '' })
+        .mockResolvedValueOnce({ stdout: '', stderr: statusAfterSwitch });
 
-      const result = await switchAccount('monalisa');
+      const result = await service.switchAccount('monalisa');
 
       expect(mockExecFileAsync).toHaveBeenNthCalledWith(1, 'gh', [
         'auth',
@@ -159,7 +163,7 @@ describe('auth service', () => {
     it.each(['', '-invalid', 'user--name', 'a'.repeat(40)])(
       'should reject invalid username: %s',
       async (name) => {
-        await expect(switchAccount(name)).rejects.toMatchObject({
+        await expect(service.switchAccount(name)).rejects.toMatchObject({
           message: 'Invalid GitHub username',
           statusCode: 400,
         });
@@ -171,7 +175,7 @@ describe('auth service', () => {
         new Error('account "unknown" not found')
       );
 
-      const result = switchAccount('unknown');
+      const result = service.switchAccount('unknown');
       await expect(result).rejects.toThrow(
         'Failed to switch GitHub account to "unknown"'
       );
